// src/app/types.tsx
export type ExRate = {
    result: number;
    cur_unit: string;
    ttb: string;
    tts: string;
    deal_bas_r: string;
    bkpr: string;
    yy_efee_r: string;
    ten_dd_efee_r: string;
    kftc_bkpr: string;
    kftc_deal_bas_r: string;
    cur_nm: string;
};
export interface ExchangeProps {
    data: ExRate[];
    // selectedCurrency: string;
    // onCurrencyChange: (event: React.ChangeEvent<HTMLSelectElement>) => void;
}